% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodebb_setup.R
\name{nodebb_setup}
\alias{nodebb_setup}
\title{Setup NodeBB API}
\usage{
nodebb_setup(url = NULL, json_token = NULL, bearer_token = NULL)
}
\arguments{
\item{url}{NodeBB base url (e.g., 'https://domain.com')}

\item{json_token}{JSON Web Token Secret}

\item{bearer_token}{User token}
}
\description{
\code{nodebb_setup} sets the base URL and API token for the environment.
}

