% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodebb_topic_post.R
\name{nodebb_topic_post}
\alias{nodebb_topic_post}
\title{Post new topic}
\usage{
nodebb_topic_post(cid = NULL, content = NULL, title = NULL,
  url = nodebb_url, token = nodebb_token)
}
\arguments{
\item{cid}{Category ID}

\item{content}{String}

\item{title}{Topic title}
}
\description{
\code{nodebb_topic_posts} posts a new topic to a given category.
}

