% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodebb_topic_reply.R
\name{nodebb_topic_reply}
\alias{nodebb_topic_reply}
\title{Post topic reply}
\usage{
nodebb_topic_reply(tid = NULL, content = NULL, url = nodebb_url,
  token = nodebb_token)
}
\arguments{
\item{tid}{Topic ID}

\item{content}{String}
}
\description{
\code{nodebb_topic_reply} replies to an existing topic.
}

