% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodebbAPI.R
\docType{package}
\name{nodebbAPI}
\alias{nodebbAPI}
\alias{nodebbAPI-package}
\title{NodeBB API Interface}
\description{
nodebb_api exposes various endpoints from the nodebb-plugin-write-api plugin.
}
\author{
Jason A. French
}
\references{
\url{https://github.com/NodeBB/nodebb-plugin-write-api}
}

